refactor(ui): move run command and backup helpers to internal/ui (#18)
