+++ internal/ui/runner.go
@@ -1,4 +1,4 @@
-package system
+package ui
 
 import (
 	"bufio"
