+++ main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/mdgspace/sysreplicate/system"
+	"github.com/mdgspace/sysreplicate/internal/ui"
 )
 // main is the entry point for the program.
 func main() {
-	system.Run()
+	ui.Run()
 }
